refactor: update imports and improve typeguards
