+++ src/components/viewers/docs-viewer/schema-type/SchemaType.tsx
@@ -3,10 +3,10 @@ import { Context } from '../../../../contexts';
 import { LangContext } from '../../../../contexts/types';
 
 import { useAppSelector } from '../../../../store/hooks';
-import { Docs } from '../../../../store/reducers/docs-slice';
 import { AllTypes, SchemaTypeProps } from '../../../../store/types';
 import { getType } from '../../../../utils/schema-resolvers';
 import SchemaItem from '../schema-item';
+import { Docs } from '../../../../store/reducers';
 
 const SchemaType: React.FC<SchemaTypeProps> = ({ data }): JSX.Element => {
   const context: LangContext = useContext<LangContext>(Context);
@@ -17,15 +17,16 @@ const SchemaType: React.FC<SchemaTypeProps> = ({ data }): JSX.Element => {
   const mainData: AllTypes = useAppSelector(Docs.mainData.select);
   const type = getType(data.type, mainData);
 
-  if (type)
-    return (
-      <>
-        <h6 className="mt-3">{typeHeader}</h6>
-        <SchemaItem data={type} />
-      </>
-    );
-
-  return <></>;
+  return (
+    <>
+      {type && (
+        <>
+          <h6 className="mt-3">{typeHeader}</h6>
+          <SchemaItem data={type} />
+        </>
+      )}
+    </>
+  );
 };
 
 export default SchemaType;
