+++ src/store/reducers/index.ts
@@ -1,19 +1,24 @@
 import { Options } from './options-slice';
 import { Inputs } from './inputs-slice';
 import { Visibility } from './visibility-slice';
+import { Docs } from './docs-slice';
 import messageReducer from './message-slice';
 import optionsReducer from './options-slice';
 import inputsReducer from './inputs-slice';
 import visibilityReducer from './visibility-slice';
 import docsReducer from './docs-slice';
+import { useGetDataQuery, useGetSchemaQuery } from './api-slice';
 
 export {
   Options,
   Inputs,
   Visibility,
+  Docs,
   messageReducer,
   optionsReducer,
   inputsReducer,
   visibilityReducer,
   docsReducer,
+  useGetDataQuery,
+  useGetSchemaQuery,
 };
